chore: Plugin and PluginResource are now factories of DependentResources
This is due to the fact that we need to be able to return now a single type of
DependentResources but several (e.g. Role, RoleBinding, etc.) for a given
capability. Client is providing the requested type which is then used by the
server to figure out which dependent resource to use to fulfill the call in
a demultiplexing-like fashion. This allows to avoid having to have a server
per DependentResource