fix: plugin needs to return all dependents
This is needed because dependents need to be created. If we
only return watched dependents, we miss potentially others
that won't be created (that was the case here for secret, role
and role binding). Need to clean up the interfaces as currently
ReadyFor is used for that but it's not intuitive and the flow
can probably be cleaned up.